pyproject.toml
[tool.poetry]
name = "external-api-weather-service"
version = "0.1.0"
description = ""
authors = ["daniskazan <danis.sharafiev@bk.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.90.0"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
backoff = "^2.2.1"
aiohttp = "^3.8.3"
pyhumps = "^3.8.0"

[tool.poetry.dev-dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

